use new sql-common test format
