+++ .travis.yml
@@ -10,7 +10,7 @@ addons:
 
 before_install:
     - export CXX="g++-4.8"
-    - npm install juttle@^0.4.0
+    - npm install juttle@>=0.4.0
 
 node_js:
     - '4.2'
