+++ package.json
@@ -18,7 +18,7 @@
     "lint": "npm run jshint && npm run style"
   },
   "dependencies": {
-    "juttle-sql-adapter-common": "^0.3.0",
+    "juttle-sql-adapter-common": "^0.3.1",
     "knex": "^0.9.0",
     "pg": "^4.4.3",
     "underscore": "^1.8.3"
