+++ test/db.spec.js
@@ -1,0 +1,2 @@
+require('juttle-sql-adapter-common/test/utils');
+require('juttle-sql-adapter-common/test/db.spec');
