+++ test/filters.spec.js
@@ -1,0 +1,2 @@
+require('./conf');
+require('juttle-sql-adapter-common/test/filters.spec');
