+++ test/mocha.opts
@@ -1,1 +1,1 @@
---timeout 5000
+--timeout 10000
